refactor(runtime): move workspace copy executor
