+++ crates/extensions/ironclaw_extension_support/Cargo.toml
@@ -34,6 +34,7 @@ blake3 = "1"
 chrono = { version = "0.4", features = ["serde"] }
 futures-util = "0.3"
 glob = "0.3"
+hex = "0.4"
 ironclaw_auth = { path = "../../domains/ironclaw_auth" }
 ironclaw_extractors = { path = "../../domains/ironclaw_extractors" }
 ironclaw_filesystem = { path = "../../substrates/ironclaw_filesystem" }
@@ -44,6 +45,7 @@ ironclaw_skills = { path = "../../domains/ironclaw_skills" }
 regex = "1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+sha2 = "0.11"
 similar = "3"
 thiserror = "2"
 tokio = { version = "1", features = ["rt", "sync"] }
